+++ cmd/abci-cli/abci-cli.go
@@ -92,6 +92,7 @@ type response struct {
 	// generic abci response
 	Data []byte
 	Code uint32
+	Info string
 	Log  string
 
 	Query *queryResponse
@@ -184,7 +185,7 @@ var consoleCmd = &cobra.Command{
 	Use:   "console",
 	Short: "start an interactive ABCI console for multiple commands",
 	Long: `start an interactive ABCI console for multiple commands
-	
+
 This command opens an interactive console for running any of the other commands
 without opening a new connection each time
 `,
@@ -519,14 +520,11 @@ func cmdSetOption(cmd *cobra.Command, args []string) error {
 	}
 
 	key, val := args[0], args[1]
-	res, err := client.SetOptionSync(types.RequestSetOption{key, val})
+	_, err := client.SetOptionSync(types.RequestSetOption{key, val})
 	if err != nil {
 		return err
 	}
-	printResponse(cmd, args, response{
-		Code: res.Code,
-		Log:  res.Log,
-	})
+	printResponse(cmd, args, response{Log: "OK (SetOption doesn't return anything.)"}) // NOTE: Nothing to show...
 	return nil
 }
 
@@ -550,6 +548,7 @@ func cmdDeliverTx(cmd *cobra.Command, args []string) error {
 	printResponse(cmd, args, response{
 		Code: res.Code,
 		Data: res.Data,
+		Info: res.Info,
 		Log:  res.Log,
 	})
 	return nil
@@ -560,7 +559,7 @@ func cmdCheckTx(cmd *cobra.Command, args []string) error {
 	if len(args) == 0 {
 		printResponse(cmd, args, response{
 			Code: codeBad,
-			Log:  "want the tx",
+			Info: "want the tx",
 		})
 		return nil
 	}
@@ -575,6 +574,7 @@ func cmdCheckTx(cmd *cobra.Command, args []string) error {
 	printResponse(cmd, args, response{
 		Code: res.Code,
 		Data: res.Data,
+		Info: res.Info,
 		Log:  res.Log,
 	})
 	return nil
@@ -587,9 +587,7 @@ func cmdCommit(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	printResponse(cmd, args, response{
-		Code: res.Code,
 		Data: res.Data,
-		Log:  res.Log,
 	})
 	return nil
 }
@@ -599,7 +597,8 @@ func cmdQuery(cmd *cobra.Command, args []string) error {
 	if len(args) == 0 {
 		printResponse(cmd, args, response{
 			Code: codeBad,
-			Log:  "want the query",
+			Info: "want the query",
+			Log:  "",
 		})
 		return nil
 	}
@@ -619,6 +618,7 @@ func cmdQuery(cmd *cobra.Command, args []string) error {
 	}
 	printResponse(cmd, args, response{
 		Code: resQuery.Code,
+		Info: resQuery.Info,
 		Log:  resQuery.Log,
 		Query: &queryResponse{
 			Key:    resQuery.Key,