+++ types/result.go
@@ -3,7 +3,6 @@ package types
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 
 	"github.com/gogo/protobuf/jsonpb"
 )
@@ -22,11 +21,6 @@ func (r ResponseCheckTx) IsErr() bool {
 	return r.Code != CodeTypeOK
 }
 
-// Error implements error interface by formatting response as string.
-func (r ResponseCheckTx) Error() string {
-	return fmtError(r.Code, r.Log)
-}
-
 // IsOK returns true if Code is OK.
 func (r ResponseDeliverTx) IsOK() bool {
 	return r.Code == CodeTypeOK
@@ -35,26 +29,6 @@ func (r ResponseDeliverTx) IsOK() bool {
 // IsErr returns true if Code is something other than OK.
 func (r ResponseDeliverTx) IsErr() bool {
 	return r.Code != CodeTypeOK
-}
-
-// Error implements error interface by formatting response as string.
-func (r ResponseDeliverTx) Error() string {
-	return fmtError(r.Code, r.Log)
-}
-
-// IsOK returns true if Code is OK.
-func (r ResponseCommit) IsOK() bool {
-	return r.Code == CodeTypeOK
-}
-
-// IsErr returns true if Code is something other than OK.
-func (r ResponseCommit) IsErr() bool {
-	return r.Code != CodeTypeOK
-}
-
-// Error implements error interface by formatting response as string.
-func (r ResponseCommit) Error() string {
-	return fmtError(r.Code, r.Log)
 }
 
 // IsOK returns true if Code is OK.
@@ -67,15 +41,6 @@ func (r ResponseQuery) IsErr() bool {
 	return r.Code != CodeTypeOK
 }
 
-// Error implements error interface by formatting response as string.
-func (r ResponseQuery) Error() string {
-	return fmtError(r.Code, r.Log)
-}
-
-func fmtError(code uint32, log string) string {
-	return fmt.Sprintf("Error code (%d): %s", code, log)
-}
-
 //---------------------------------------------------------------------------
 // override JSON marshalling so we dont emit defaults (ie. disable omitempty)
 // note we need Unmarshal functions too because protobuf had the bright idea
@@ -84,7 +49,7 @@ func fmtError(code uint32, log string) string {
 var (
 	jsonpbMarshaller = jsonpb.Marshaler{
 		EnumsAsInts:  true,
-		EmitDefaults: true,
+		EmitDefaults: false,
 	}
 	jsonpbUnmarshaller = jsonpb.Unmarshaler{}
 )